Create SCP Directory & DB Backup Script
This script backups a remote directory and mysql database to a local machine. The script begins by logging into the remote server compressing the target directory, creating a mysql dump, taking checksums of the files, and downloading all of the target files and dumps to the local machine, then verifying the checksums. For automated backups one needs to manually create a cron job.